with prettyurls false

+++ src/get_vaccine_stock_coverage.jl
@@ -1,6 +1,6 @@
 """
     get_vaccine_stock_coverage(k, parameters)
-    
+
 Returns el percentage of popullation to vaccine when the inventory 
 level of interest is k and use the current parameters 
 
@@ -9,7 +9,7 @@ level of interest is k and use the current parameters
 - `parameters::DataFrame`: current parameters.
 ...
 """
-function get_vaccine_stock_coverage(k::Float64, parameters)
+function get_vaccine_stock_coverage(k, parameters)
     l_s = parameters.low_stock[1] / parameters.N[1]
     x_coverage = maximum([k - l_s, 0.0])
     return x_coverage
